requirements.txt
bson==0.5.6
Django==2.0.7
gunicorn==19.9.0
pymodm==0.4.0
pymongo==3.6.1
python-dotenv==0.8.2
djangorestframework==3.9.2
coverage==4.5.1